requirements.txt
argcomplete==3.6.0
catkin-pkg==1.0.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
conan2==0.0.4
coverage==7.6.12
distlib==0.3.9
docutils==0.21.2
empy==3.3.4
iniconfig==2.0.0
lark==1.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.2.4
packaging==24.2
pluggy==1.5.0
psutil==7.0.0
Pygments==2.19.1
pyparsing==3.2.1
pyrootutils==1.0.4
pytest==8.3.5
pytest-cov==6.0.0
pytest-repeat==0.9.3
pytest-rerunfailures==15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
rich==13.9.4
rpyutils==0.1.4
setuptools==76.0.0
six==1.17.0
tqdm==4.67.1
vcstool==0.3.0
